% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_scores.R
\name{compare_scores}
\alias{compare_scores}
\title{Compare Multiple Scoring Methods}
\usage{
compare_scores(outcome, scores, bins = 10)
}
\arguments{
\item{outcome}{Vector of outcomes}

\item{scores}{Named list of score vectors}

\item{bins}{Number of bins}
}
\value{
Data frame with PII and RMSE
}
\description{
Compare Multiple Scoring Methods
}
