% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_batch.R
\name{pii_batch}
\alias{pii_batch}
\title{Compute PII for Multiple Scores}
\usage{
pii_batch(formula, data, bins = 10)
}
\arguments{
\item{formula}{A formula with multiple predictors (e.g., outcome ~ x1 + x2 + x3)}

\item{data}{A data.frame}

\item{bins}{Number of bins for discretization}
}
\value{
A data.frame of PII and RMSE for each predictor
}
\description{
Evaluates PII and RMSE for multiple predictors
}
