% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_plot.R
\name{pii_plot}
\alias{pii_plot}
\title{Plot Score vs Outcome for Visual PII Insight}
\usage{
pii_plot(score, outcome, bins = 10)
}
\arguments{
\item{score}{Numeric score vector}

\item{outcome}{Numeric outcome vector}

\item{bins}{Number of bins for discretization}
}
\value{
A base R plot
}
\description{
Plot Score vs Outcome for Visual PII Insight
}
