% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_test.R
\name{pii_test}
\alias{pii_test}
\title{Permutation Test for PII}
\usage{
pii_test(score, outcome, bins = 10, n_perm = 1000)
}
\arguments{
\item{score}{Numeric score vector}

\item{outcome}{Outcome vector}

\item{bins}{Number of bins for discretization}

\item{n_perm}{Number of permutations}
}
\value{
List with observed PII, p-value, and null distribution
}
\description{
Permutation Test for PII
}
